% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_set.R
\name{check}
\alias{check}
\title{Perform a test, returning the score and a message}
\usage{
check(match = TRUE, expr, message = NULL, pts = 1)
}
\description{
Perform a test, returning the score and a message
}

