% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-253.R
\name{score253}
\alias{score253}
\title{Score a daily activity in Math 253}
\usage{
score253(day)
}
\arguments{
\item{day.}{A single number, e.g. 1, 2, 17. The day of the activity.  This identifies which test
statements are to be used.}
}
\description{
This function is specifically designed for the programming activities in Math 253
}

